requirements.lock.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
bitarray==3.8.0
certifi==2025.11.12
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.8.0
hexbytes==1.3.1
idna==3.11
iniconfig==2.3.0
multidict==6.7.0
packaging==25.0
parsimonious==0.10.0
pluggy==1.6.0
propcache==0.4.1
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
pyunormalize==17.0.0
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
toolz==1.1.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
web3==7.14.0
websockets==15.0.1
yarl==1.22.0
